feat(web): carry the app header into settings and move sign-out into it
The settings route kept only its own chrome, so there was no way back to
Agents or Analytics from it — worst for a non-admin, who landed on the
no-access screen with no navigation at all. Mount AppHeader above the
settings SPA, outside its Suspense boundary, so the bar is present in every
state of the route: auth loading, session error, no-access and the settings
content itself.

Make the header's right-hand actions legible. The Settings control was a
bare muted gear that read as decoration; it now uses the same icon+label
treatment and active pill as the primary nav, since it is a destination
like the others. Sign out moves out of the sidebar footer to sit beside it,
separated by a divider and carrying slightly quieter ink — it ends the
session rather than navigating, and that difference should be visible.
Sign-out behaviour is unchanged: best-effort POST to /api/auth/logout, then
clear local state and return to /login.

Settings stays admin-only and still defaults to hidden until /api/auth/me
resolves; sign out shows for everyone.
