+++ web/components/sidebar.tsx
@@ -1,13 +1,12 @@
 "use client"
 
-import { Search, Pin, X, ChevronDown, PanelLeftClose, PanelLeft, Loader2, AlertCircle, LogOut, Plus } from "lucide-react"
+import { Search, Pin, X, ChevronDown, PanelLeftClose, PanelLeft, Loader2, AlertCircle, Plus } from "lucide-react"
 import { useBranding } from "@/hooks/use-branding"
 import { Button } from "@/components/ui/button"
 import { Input } from "@/components/ui/input"
 import { ScrollArea } from "@/components/ui/scroll-area"
 import { ClawCard } from "@/components/claw-card"
-import { clearConfig, fetchWorkspaces, type Workflow } from "@/lib/api"
-import { getAuthToken } from "@/lib/auth-storage"
+import { fetchWorkspaces, type Workflow } from "@/lib/api"
 import {
   DropdownMenu,
   DropdownMenuContent,
@@ -470,31 +469,6 @@ export function Sidebar({
           ) : null}
         </DragOverlay>
       </DndContext>
-
-      {/* Logout */}
-      <div className="p-2 border-t border-border">
-        <div className="flex items-center gap-1">
-          <Button
-            variant="ghost"
-            size="sm"
-            className="flex-1 justify-start gap-2 text-muted-foreground hover:text-foreground"
-            onClick={async () => {
-              const token = getAuthToken() || ""
-              if (token) {
-                const { getHubUrl } = await import("@/lib/hub-url")
-                const hubUrl = getHubUrl()
-                const logoutUrl = hubUrl ? `${hubUrl}/api/auth/logout` : "/api/auth/logout"
-                await fetch(logoutUrl, { method: "POST", headers: { Authorization: `Bearer ${token}` } }).catch(() => {})
-              }
-              clearConfig()
-              window.location.href = "/login"
-            }}
-          >
-            <LogOut className="size-4" />
-            Sign out
-          </Button>
-        </div>
-      </div>
     </aside>
     )
   }
